sw floattable: fix outdated text frame portions after split
The document has a floating table of 2 pages, with some anchor text on
page 2. Split of the anchor text resulted in 2 paragraphs: the first
still hosts the floating table and now lacks a fly portion (overlap of
text and floating table) and the second still wraps around the floating
table but has the correct position.

What happens is that we format the first text frame, then its flys
(which moves the follow fly from page 1 to page 2) and then we format
the second text frame. This means that the first text frame should have
a fly portion to wrap around the floating table, but it does not have,
because SwTextFrame::Format() is called for the first text frame, then
SwPageFrame::MoveFly() is called, then SwTextFrame::Format() for the
second text frame, i.e. no reformat of the first text frame after moving
the fly.

Fix the problem somewhat similar to what commit
cf2c070de2bafeec3b476c6bff7bb4ac87ba46db (sw layout: invalidate margins
of body content when moving a fly from page, 2022-12-09) did, but here
what's outdated is the portions of the text frame, and only for the
anchor text frame.

This fixes the problem in practice, but note that in theory this could
affect non-split flys and also other, non-anchor text frames as well.
Limit the update of the portion list to the anchor of the split fly, but
we may want to do the same for all lowers of the body frame if there is
a practical need for it.

(cherry picked from commit 5fec60b4732bdbdb681be08e43a9be47c3bfb320)

Change-Id: I3067e78d30e16ed3d2f02a80cff4cd84d9bdcf3b
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/160819
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
Reviewed-by: Caolán McNamara <caolan.mcnamara@collabora.com>